+++ deploy/kubectl_helpers/shell.sh
@@ -8,9 +8,4 @@ COMPONENT=$2
 
 POD_NAME=$("${DIR}"/utils/get_pod_name.sh "$@")
 
-case ${COMPONENT} in
-  seqr) CONTAINER=seqr-pod ;;
-  *) CONTAINER=${COMPONENT} ;;
-esac
-
-kubectl exec -it "${POD_NAME}" -c "${CONTAINER}" -- /bin/bash
+kubectl exec -it "${POD_NAME}" -c "${COMPONENT}" -- /bin/bash