+++ reference_data/management/commands/utils/update_utils.py
@@ -4,6 +4,8 @@ import gzip
 from tqdm import tqdm
 import traceback
 from django.core.management.base import BaseCommand, CommandError
+from django.db import models
+
 from reference_data.management.commands.utils.download_utils import download_file
 from reference_data.management.commands.utils.gene_utils import get_genes_by_symbol_and_id
 from reference_data.models import GeneInfo
@@ -13,7 +15,7 @@ logger = logging.getLogger(__name__)
 
 class ReferenceDataHandler(object):
 
-    model_cls = None
+    model_cls = models.Model
     url = None
     header_fields = None
     post_process_models = None
@@ -55,6 +57,61 @@ class ReferenceDataHandler(object):
             raise ValueError('Gene "{}" not found in the GeneInfo table'.format(gene_id or gene_symbol))
         return gene
 
+    def update_records(self, file_path=None):
+        """
+        Args:
+            file_path (str): optional local file path. If not specified, or the path doesn't exist, the table will be downloaded.
+        """
+        model_name = self.model_cls.__name__
+        model_objects = getattr(self.model_cls, 'objects')
+
+        logger.info(f'Updating {model_name}')
+
+        if not file_path or not os.path.isfile(file_path):
+            file_path = download_file(self.url)
+
+        models = []
+        skip_counter = 0
+        logger.info('Parsing file')
+        open_file = gzip.open if file_path.endswith('.gz') else open
+        open_mode = 'rt' if file_path.endswith('.gz') else 'r'
+        try:
+            with open_file(file_path, open_mode) as f:
+                header_fields = self.get_file_header(f)
+
+                for line in self.get_file_iterator(f):
+                    record = dict(zip(header_fields, line if isinstance(line, list) else line.rstrip('\r\n').split('\t')))
+                    for record in self.parse_record(record):
+                        if record is None:
+                            continue
+
+                        try:
+                            record[self.gene_key] = self.get_gene_for_record(record)
+                        except ValueError as e:
+                            skip_counter += 1
+                            logger.debug(e)
+                            continue
+
+                        models.append(self.model_cls(**record))
+
+            if self.post_process_models is not None:
+                self.post_process_models(models)
+
+            if not self.keep_existing_records:
+                logger.info("Deleting {} existing {} records".format(model_objects.count(), model_name))
+                model_objects.all().delete()
+
+            logger.info("Creating {} {} records".format(len(models), model_name))
+            model_objects.bulk_create(models)
+
+            logger.info("Done")
+            logger.info("Loaded {} {} records from {}. Skipped {} records with unrecognized genes.".format(
+                model_objects.count(), model_name, file_path, skip_counter))
+            if skip_counter > 0:
+                logger.info('Running ./manage.py update_gencode to update the gencode version might fix missing genes')
+        except Exception as e:
+            logger.error(str(e), extra={'traceback': traceback.format_exc()})
+
 
 class GeneCommand(BaseCommand):
     reference_data_handler = ReferenceDataHandler
@@ -65,61 +122,4 @@ class GeneCommand(BaseCommand):
                             default=os.path.join('resource_bundle', os.path.basename(self.reference_data_handler.url)))
 
     def handle(self, *args, **options):
-        update_records(self.reference_data_handler(**options), file_path=options.get('file_path'), )
-
-
-def update_records(reference_data_handler, file_path=None):
-    """
-    Args:
-        file_path (str): optional local file path. If not specified, or the path doesn't exist, the table will be downloaded.
-    """
-    model_cls = reference_data_handler.model_cls
-    model_name = model_cls.__name__
-    model_objects = getattr(model_cls, 'objects')
-
-    logger.info(f'Updating {model_name}')
-
-    if not file_path or not os.path.isfile(file_path):
-        file_path = download_file(reference_data_handler.url)
-
-    models = []
-    skip_counter = 0
-    logger.info('Parsing file')
-    open_file = gzip.open if file_path.endswith('.gz') else open
-    open_mode = 'rt' if file_path.endswith('.gz') else 'r'
-    try:
-        with open_file(file_path, open_mode) as f:
-            header_fields = reference_data_handler.get_file_header(f)
-
-            for line in reference_data_handler.get_file_iterator(f):
-                record = dict(zip(header_fields, line if isinstance(line, list) else line.rstrip('\r\n').split('\t')))
-                for record in reference_data_handler.parse_record(record):
-                    if record is None:
-                        continue
-
-                    try:
-                        record[reference_data_handler.gene_key] = reference_data_handler.get_gene_for_record(record)
-                    except ValueError as e:
-                        skip_counter += 1
-                        logger.debug(e)
-                        continue
-
-                    models.append(model_cls(**record))
-
-        if reference_data_handler.post_process_models:
-            reference_data_handler.post_process_models(models)
-
-        if not reference_data_handler.keep_existing_records:
-            logger.info("Deleting {} existing {} records".format(model_objects.count(), model_name))
-            model_objects.all().delete()
-
-        logger.info("Creating {} {} records".format(len(models), model_name))
-        model_objects.bulk_create(models)
-
-        logger.info("Done")
-        logger.info("Loaded {} {} records from {}. Skipped {} records with unrecognized genes.".format(
-            model_objects.count(), model_name, file_path, skip_counter))
-        if skip_counter > 0:
-            logger.info('Running ./manage.py update_gencode to update the gencode version might fix missing genes')
-    except Exception as e:
-        logger.error(str(e), extra={'traceback': traceback.format_exc()})
+        self.reference_data_handler(**options).update_records(file_path=options.get('file_path'))