+++ reference_data/management/tests/update_all_reference_data_tests.py
@@ -4,6 +4,14 @@ from django.core.management import call_command
 from django.core.management.base import CommandError
 from django.test import TestCase
 
+from reference_data.management.commands.utils.update_utils import ReferenceDataHandler
+from reference_data.management.commands.update_dbnsfp_gene import DbNSFPReferenceDataHandler
+from reference_data.management.commands.update_gene_constraint import GeneConstraintReferenceDataHandler
+from reference_data.management.commands.update_gene_cn_sensitivity import CNSensitivityReferenceDataHandler
+from reference_data.management.commands.update_gencc import GenCCReferenceDataHandler
+from reference_data.management.commands.update_clingen import ClinGenReferenceDataHandler
+from reference_data.management.commands.update_refseq import RefseqReferenceDataHandler
+
 
 def omim_exception(omim_key):
     raise Exception('Omim exception, key: '+omim_key)
@@ -26,22 +34,14 @@ class UpdateAllReferenceDataTest(TestCase):
     fixtures = ['users', 'reference_data']
 
     def setUp(self):
-        patcher = mock.patch('reference_data.management.commands.update_dbnsfp_gene.DbNSFPReferenceDataHandler', lambda: 'dbnsfp_gene')
+        self.mock_handlers = []
+        def _mock_init_handler(_self, **kwargs):
+            self.mock_handlers.append(type(_self))
+            return None
+        patcher = mock.patch.object(ReferenceDataHandler, '__init__', _mock_init_handler)
         patcher.start()
         self.addCleanup(patcher.stop)
-        patcher = mock.patch('reference_data.management.commands.update_gene_cn_sensitivity.CNSensitivityReferenceDataHandler', lambda: 'gene_cn_sensitivity')
-        patcher.start()
-        self.addCleanup(patcher.stop)
-        patcher = mock.patch('reference_data.management.commands.update_gene_constraint.GeneConstraintReferenceDataHandler', lambda: 'gene_constraint')
-        patcher.start()
-        self.addCleanup(patcher.stop)
-        patcher = mock.patch('reference_data.management.commands.update_gencc.GenCCReferenceDataHandler', lambda: 'gencc')
-        patcher.start()
-        self.addCleanup(patcher.stop)
-        patcher = mock.patch('reference_data.management.commands.update_clingen.ClinGenReferenceDataHandler', lambda: 'clingen')
-        patcher.start()
-        self.addCleanup(patcher.stop)
-        patcher = mock.patch('reference_data.management.commands.update_refseq.RefseqReferenceDataHandler', lambda: 'refseq')
+        patcher = mock.patch.object(RefseqReferenceDataHandler, '__init__', _mock_init_handler)
         patcher.start()
         self.addCleanup(patcher.stop)
 
@@ -64,7 +64,7 @@ class UpdateAllReferenceDataTest(TestCase):
         patcher = mock.patch('reference_data.management.commands.update_all_reference_data.update_hpo')
         self.mock_update_hpo = patcher.start()
         self.addCleanup(patcher.stop)
-        patcher = mock.patch('reference_data.management.commands.update_all_reference_data.update_records')
+        patcher = mock.patch.object(ReferenceDataHandler, 'update_records')
         self.mock_update_records = patcher.start()
         self.addCleanup(patcher.stop)
         patcher = mock.patch('reference_data.management.commands.update_all_reference_data.logger')
@@ -79,7 +79,6 @@ class UpdateAllReferenceDataTest(TestCase):
         self.assertEqual(str(err.exception), 'Error: one of the arguments --omim-key --use-cached-omim --skip-omim is required')
 
         # Test update all gencode, no skips, fail primate_ai and mgi
-        self.mock_omim.return_value = 'omim'
         call_command('update_all_reference_data', '--omim-key=test_key')
 
         calls = [
@@ -94,18 +93,16 @@ class UpdateAllReferenceDataTest(TestCase):
 
         self.mock_omim.assert_called_with('test_key')
         self.mock_cached_omim.assert_not_called()
-
-        self.assertEqual(self.mock_update_records.call_count, 7)
-        calls = [
-            mock.call('omim'),
-            mock.call('dbnsfp_gene'),
-            mock.call('gene_constraint'),
-            mock.call('gene_cn_sensitivity'),
-            mock.call('gencc'),
-            mock.call('clingen'),
-            mock.call('refseq'),
-        ]
-        self.mock_update_records.assert_has_calls(calls)
+        self.mock_omim.return_value.update_records.assert_called_with()
+        self.assertEqual(self.mock_update_records.call_count, 6)
+        self.assertListEqual(self.mock_handlers, [
+            DbNSFPReferenceDataHandler,
+            GeneConstraintReferenceDataHandler,
+            CNSensitivityReferenceDataHandler,
+            GenCCReferenceDataHandler,
+            ClinGenReferenceDataHandler,
+            RefseqReferenceDataHandler,
+        ])
 
         self.mock_update_hpo.assert_called_with()
 
@@ -134,13 +131,12 @@ class UpdateAllReferenceDataTest(TestCase):
         self.mock_logger.info.assert_called_with("Done")
 
     def test_cached_omim_update_reference_data_command(self):
-        self.mock_cached_omim.return_value = 'cached_omim'
-
         call_command(
             'update_all_reference_data', '--use-cached-omim', *SKIP_ARGS)
 
         self.mock_cached_omim.assert_called_with()
-        self.mock_update_records.assert_called_with('cached_omim')
+        self.mock_cached_omim.return_value.update_records.assert_called_with()
+        self.mock_update_records.assert_not_called()
 
         self.mock_omim.assert_not_called()
         self.mock_update_gencode.assert_not_called()