+++ seqr/management/commands/check_for_new_samples_from_pipeline.py
@@ -14,7 +14,7 @@ from seqr.utils.file_utils import file_iter, list_files
 from seqr.utils.search.add_data_utils import notify_search_data_loaded
 from seqr.utils.search.utils import parse_valid_variant_id
 from seqr.utils.search.hail_search_utils import hail_variant_multi_lookup, search_data_type
-from seqr.views.utils.airtable_utils import AirtableSession, LOADABLE_PDO_STATUSES, AVAILABLE_PDO_STATUS
+from seqr.views.utils.airtable_utils import AirtableSession, LOADABLE_PDO_STATUSES, AVAILABLE_PDO_STATUS, LOADING_PDO_STATUS
 from seqr.views.utils.dataset_utils import match_and_update_search_samples
 from seqr.views.utils.permissions_utils import is_internal_anvil_project, project_has_anvil
 from seqr.views.utils.variant_utils import reset_cached_search_results, update_projects_saved_variant_json, \
@@ -29,6 +29,7 @@ RUN_PATH_FIELDS = ['genome_version', 'dataset_type', 'run_version']
 DATASET_TYPE_MAP = {'GCNV': Sample.DATASET_TYPE_SV_CALLS}
 USER_EMAIL = 'manage_command'
 MAX_LOOKUP_VARIANTS = 5000
+RELATEDNESS_CHECK_NAME = 'relatedness_check'
 
 PDO_COPY_FIELDS = [
     'PDO', 'PDOStatus', 'SeqrLoadingDate', 'GATKShortReadCallsetPath', 'SeqrProjectURL', 'TerraProjectURL',
@@ -160,6 +161,7 @@ class Command(BaseCommand):
                 split_project_pdos[project.name] = cls._update_pdos(session, project.guid, sample_ids)
 
         # Send failure notifications
+        relatedness_check_file_path = metadata.get('relatedness_check_file_path')
         failed_family_samples = metadata.get('failed_family_samples', {})
         failed_families_by_guid = {f['guid']: f for f in Family.objects.filter(
             guid__in={family for families in failed_family_samples.values() for family in families}
@@ -176,9 +178,15 @@ class Command(BaseCommand):
                 split_pdos = split_project_pdos.get(project)
                 if split_pdos:
                     summary += f'\n\nSkipped samples in this project have been moved to {", ".join(split_pdos)}'
+
+                relatedness_check_message = (
+                    f'\nRelatedness check results: {relatedness_check_file_path}'
+                    if (relatedness_check_file_path and check == RELATEDNESS_CHECK_NAME)
+                    else ''
+                )
                 safe_post_to_slack(
                     SEQR_SLACK_LOADING_NOTIFICATION_CHANNEL,
-                    f'The following {len(failures)} families failed {check.replace("_", " ")} in {project}:\n{summary}'
+                    f'The following {len(failures)} families failed {check.replace("_", " ")} in {project}:\n{summary}{relatedness_check_message}'
                 )
 
         # Reload saved variant JSON
@@ -220,7 +228,7 @@ class Command(BaseCommand):
         # Create PDOs and then update Samples with new PDOs
         # Does not create PDOs with Samples directly as that would not remove Samples from old PDOs
         new_pdos = session.safe_create_records('PDO', [
-            {'PDO': pdo_name, **pdo} for pdo_name, (_, pdo) in pdos_to_create.items()
+            {'PDO': pdo_name, **pdo, 'PDOStatus': LOADING_PDO_STATUS} for pdo_name, (_, pdo) in pdos_to_create.items()
         ])
         pdo_id_map = {pdos_to_create[record['fields']['PDO']][0]: record['id'] for record in new_pdos}
         for pdo_id, sample_record_ids in skipped_pdo_samples.items():