+++ ui/pages/Project/components/edit-families-and-individuals/EditIndividualsForm.jsx
@@ -2,6 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
 
+import { getUser } from 'redux/selectors'
 import DataLoader from 'shared/components/DataLoader'
 import EditRecordsForm from 'shared/components/form/EditRecordsForm'
 import { FAMILY_FIELD_ID, INDIVIDUAL_FIELD_ID } from 'shared/utils/constants'
@@ -9,14 +10,16 @@ import { INDIVIDUAL_FIELDS } from '../../constants'
 import { loadIndividuals, updateIndividuals } from '../../reducers'
 import { getProjectAnalysisGroupIndividualsByGuid, getIndivdualsLoading } from '../../selectors'
 
-const EditIndividualsForm = React.memo(({ load, loading, ...props }) => (
+const INDIVIDUAL_CORE_FIELDS = INDIVIDUAL_FIELDS.slice(0, -1)
+
+const EditIndividualsForm = React.memo(({ load, loading, user, ...props }) => (
   <DataLoader load={load} content={props.records} loading={loading}>
     <EditRecordsForm
       idField="individualGuid"
       entityKey="individuals"
       defaultSortColumn={FAMILY_FIELD_ID}
       filterColumn={INDIVIDUAL_FIELD_ID}
-      columns={INDIVIDUAL_FIELDS}
+      columns={user.isAnalyst ? INDIVIDUAL_FIELDS : INDIVIDUAL_CORE_FIELDS}
       {...props}
     />
   </DataLoader>
@@ -28,11 +31,13 @@ EditIndividualsForm.propTypes = {
   modalName: PropTypes.string,
   load: PropTypes.func,
   loading: PropTypes.bool,
+  user: PropTypes.object,
 }
 
 const mapStateToProps = (state, ownProps) => ({
   loading: getIndivdualsLoading(state),
   records: getProjectAnalysisGroupIndividualsByGuid(state, ownProps),
+  user: getUser(state),
 })
 
 const mapDispatchToProps = {